+++ common/performance_model/dram_perf_model.h
@@ -21,7 +21,7 @@ class DramPerfModel
    private:
       QueueModel* m_queue_model;
       MovingAverage<UInt64>* m_moving_average;
-      UInt32 m_dram_access_cost;
+      float m_dram_access_cost;
       float m_dram_bandwidth;
 
       bool m_queue_model_enabled;
